social_network/requirements.txt
asn1crypto==0.24.0
certifi==2018.11.29
chardet==3.0.4
cryptography==2.1.4
Django==2.1.5
djangorestframework==3.9.1
djangorestframework-simplejwt==3.3
idna==2.6
keyring==10.6.0
keyrings.alt==3.0
psycopg2-binary==2.7.7
pycrypto==2.6.1
pygobject==3.26.1
pyhunter==0.3
PyJWT==1.7.1
pytz==2018.9
pyxdg==0.25
requests==2.21.0
SecretStorage==2.3.1
six==1.11.0
urllib3==1.24.1
